# api/api.yaml
openapi: 3.0.0
info:
  title: ZidiBackend API
  version: 1.0.0
  description: API for managing organizations and users in ZidiBackend
paths:
...
  /users/email:
    get:
      summary: Get a user by email
      operationId: GetUserByEmail
      security:
        - BearerAuth: []
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /auth/login:
    post:
      summary: User login
      operationId: LoginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT access token
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          description: Unauthorized - Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/logout:
    post:
      summary: User logout
      operationId: LogoutUser
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Successful logout
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Successfully logged out
        "401":
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/superuser/login:
    post:
      summary: Superuser login
      operationId: SuperuserLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successful superuser login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT access token
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          description: Unauthorized - Invalid credentials or insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /campaigns/{id}:
    get:
      summary: Get a campaign by ID
      operationId: getCampaignById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Campaign found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Campaign not found

    put:
      summary: Update a campaign by ID
      operationId: updateCampaign
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Campaign"
      responses:
        "200":
          description: Campaign updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Campaign not found

    delete:
      summary: Delete a campaign by ID
      operationId: deleteCampaign
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Campaign deleted successfully
        "404":
          description: Campaign not found
...
  /balances:
    get:
      summary: Get all balances
      operationId: getAllBalances
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: List of balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Balance"
        "401":
          description: Unauthorized

    post:
      summary: Create a new balance
      operationId: createBalance
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBalanceRequest"
      responses:
        "201":
          description: Balance created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Balance"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
  /balances/{campaignId}:
    get:
      summary: Get balance by campaign
      operationId: getBalanceByCampaign
      security:
        - BearerAuth: []
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Balance retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Balance"
        "401":
          description: Unauthorized
        "404":
          description: Balance not found

    patch:
      summary: Update balance by campaign
      operationId: updateBalance
      security:
        - BearerAuth: []
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBalanceRequest"
      responses:
        "200":
          description: Balance updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Balance"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Balance not found
        "500":
          description: Internal server error
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # Optional, but helps with documentation
  schemas:
...
    Campaign:
      type: object
      required:
        - id
        - campaignName
        - couponId
        - characterType
        - couponLength
        - couponNumber
        - welcomeMessage
        - questionNumber
        - amount
        - price
        - status
        - organizationId
        - communicationChannels
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        campaignName:
          type: string
          example: "Summer Promo"
        couponId:
          type: string
          example: "Alphanumeric"
        characterType:
          type: string
          example: "Alphanumeric"
        couponLength:
          type: integer
          example: 10
        couponNumber:
          type: integer
          example: 100
        welcomeMessage:
          type: string
          example: "Welcome to our summer campaign!"
        questionNumber:
          type: integer
          example: 5
        amount:
          type: number
          format: float
          example: 100.50
        price:
          type: number
          format: float
          example: 100.50
        status:
          type: string
          example: "active"
        organizationId:
          type: string
          format: uuid
          example: "d63e4c28-f8d2-4c39-b3bb-92cd5b998a21"
        communicationChannels:
          type: string
          example: "whatsapp,telegram"
          description: "Comma-separated list of communication channels: whatsapp, telegram, or both"
        createdAt:
          type: string
          format: date-time
          nullable: true
          example: "2025-03-04T12:34:56Z"
    User:
      type: object
      required:
        - id
        - firstname
        - lastname
        - email
        - address
        - password
        - role
        - organization_id
      properties:
        id:
          type: string
          format: uuid
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
          format: email
        address:
          type: string
        password:
          type: string
          format: password
        role:
          type: string
        organization_id:
          type: string
          format: uuid
...
    Balance:
      type: object
      required:
        - id
        - campaignId
        - startingBalance
        - amount
      properties:
        id:
          type: string
          format: uuid
        campaignId:
          type: string
          format: uuid
        startingBalance:
          type: number
          format: float
        amount:
          type: number
          format: float
        campaignName:
          type: string
          description: "Name of the campaign linked to this balance"
    CreateBalanceRequest:
      type: object
      required:
        - campaignId
        - startingBalance
        - amount
      properties:
        campaignId:
          type: string
          format: uuid
        startingBalance:
          type: number
          format: float
        amount:
          type: number
          format: float
    UpdateBalanceRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: float
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
...
    Error:
      type: object
      required:
        - message
        - error_code
      properties:
        error_code:
          type: string
        message:
          type: string
  responses:
    BadRequestError:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
security:
  - BearerAuth: []